rename type and config files
