+++ server/src/trade/trade.types.ts
